bevy_ptr
Struct Ptr

pub struct Ptr<'a, A: IsAligned = Aligned>(/* private fields */);

Type-erased borrow of some unknown type chosen when constructing this type.

This type tries to act โborrow-likeโ which means that:

It should be considered immutable: its target must not be changed while this pointer is alive.

It must always points to a valid value of whatever the pointee type is.

The lifetime 'a accurately represents how long the pointer is valid for.

Must be sufficiently aligned for the unknown pointee type.

It may be helpful to think of this type as similar to &'a dyn Any but without the metadata and able to point to data that does not correspond to a Rust type.

Implementations

impl<'a> Ptr<'a, Aligned>

pub fn to_unaligned(self) -> Ptr<'a, Unaligned>

Removes the alignment requirement of this pointer

impl<A: IsAligned> Ptr<'_, A>

pub unsafe fn byte_offset(self, count: isize) -> Self

Calculates the offset from a pointer. As the pointer is type-erased, there is no size information available. The provided count parameter is in raw bytes.

See also: ptr::offset

Safety

The offset cannot make the existing ptr null, or take it out of bounds for its allocation.

If the A type parameter is Aligned then the offset must not make the resulting pointer be unaligned for the pointee type.

The value pointed by the resulting pointer must outlive the lifetime of this pointer.

pub unsafe fn byte_add(self, count: usize) -> Self

Calculates the offset from a pointer (convenience for .offset(count as isize)). As the pointer is type-erased, there is no size information available. The provided count parameter is in raw bytes.

See also: ptr::add

Safety

The offset cannot make the existing ptr null, or take it out of bounds for its allocation.

If the A type parameter is Aligned then the offset must not make the resulting pointer be unaligned for the pointee type.

The value pointed by the resulting pointer must outlive the lifetime of this pointer.

impl<'a, A: IsAligned> Ptr<'a, A>

pub unsafe fn new(inner: NonNull<u8>) -> Self

Creates a new instance from a raw pointer.

Safety

inner must point to valid value of whatever the pointee type is.

If the A type parameter is Aligned then inner must be sufficiently aligned for the pointee type.

inner must have correct provenance to allow reads of the pointee type.

The lifetime 'a must be constrained such that this Ptr will stay valid and nothing can mutate the pointee while this Ptr is live except through an UnsafeCell.

pub unsafe fn assert_unique(self) -> PtrMut<'a, A>

Transforms this Ptr into an PtrMut

Safety

Another PtrMut for the same Ptr must not be created until the first is dropped.

pub unsafe fn deref<T>(self) -> &'a T

Transforms this Ptr<T> into a &T with the same lifetime

Safety

T must be the erased pointee type for this Ptr.

If the type parameter A is Unaligned then this pointer must be sufficiently aligned for the pointee type T.

pub fn as_ptr(self) -> *mut u8

Gets the underlying pointer, erasing the associated lifetime.

If possible, it is strongly encouraged to use deref over this function, as it retains the lifetime.