Struct bevy_ptr::PtrMut

pub struct PtrMut<'a, A: IsAligned = Aligned>(_, _);

Type-erased mutable borrow of some unknown type chosen when constructing this type.

This type tries to act “borrow-like” which means that:

Pointer is considered exclusive and mutable. It cannot be cloned as this would lead to aliased mutability.

It must always points to a valid value of whatever the pointee type is.

The lifetime 'a accurately represents how long the pointer is valid for.

Must be sufficiently aligned for the unknown pointee type.

It may be helpful to think of this type as similar to &'a mut dyn Any but without the metadata and able to point to data that does not correspond to a Rust type.

Implementations

impl<'a> PtrMut<'a, Aligned>

pub fn to_unaligned(self) -> PtrMut<'a, Unaligned>

Removes the alignment requirement of this pointer

impl<A: IsAligned> PtrMut<'_, A>

pub unsafe fn byte_offset(self, count: isize) -> Self

Calculates the offset from a pointer. As the pointer is type-erased, there is no size information available. The provided count parameter is in raw bytes.

See also: ptr::offset

Safety

The offset cannot make the existing ptr null, or take it out of bounds for its allocation.

If the A type parameter is Aligned then the offset must not make the resulting pointer be unaligned for the pointee type.

pub unsafe fn byte_add(self, count: usize) -> Self

Calculates the offset from a pointer (convenience for .offset(count as isize)). As the pointer is type-erased, there is no size information available. The provided count parameter is in raw bytes.

See also: ptr::add

Safety

The offset cannot make the existing ptr null, or take it out of bounds for its allocation.

If the A type parameter is Aligned then the offset must not make the resulting pointer be unaligned for the pointee type.

impl<'a, A: IsAligned> PtrMut<'a, A>

pub unsafe fn new(inner: NonNull<u8>) -> Self

Creates a new instance from a raw pointer.

Safety

inner must point to valid value of whatever the pointee type is.

If the A type parameter is Aligned then inner must be sufficiently aligned for the pointee type.

inner must have correct provenance to allow read and writes of the pointee type.

The lifetime 'a must be constrained such that this PtrMut will stay valid and nothing else can read or mutate the pointee while this PtrMut is live.

pub unsafe fn promote(self) -> OwningPtr<'a, A>

Transforms this PtrMut into an OwningPtr

Safety

Must have right to drop or move out of PtrMut.

pub unsafe fn deref_mut<T>(self) -> &'a mut T

Transforms this PtrMut<T> into a &mut T with the same lifetime

Safety

T must be the erased pointee type for this PtrMut.

If the type parameter A is Unaligned then this pointer must be sufficiently aligned for the pointee type T.

pub fn as_ptr(&self) -> *mut u8

Gets the underlying pointer, erasing the associated lifetime.

If possible, it is strongly encouraged to use deref_mut over this function, as it retains the lifetime.

pub fn reborrow(&mut self) -> PtrMut<'_, A>

Gets a PtrMut from this with a smaller lifetime.

pub fn as_ref(&self) -> Ptr<'_, A>

Gets an immutable reference from this mutable reference