bevy_ptr

Trait IsAligned

pub trait IsAligned: Sealed { }

Trait that is only implemented for Aligned and Unaligned to work around the lack of ability to have const generics of an enum.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.